refactor(db): replace cache package with new db cache helpers and redis cache
- Remove cache package implementation and move caching helpers to db package
- Add new cache-helpers.ts in db package for caching logic
- Add redis-cache.ts implementing Drizzle ORM cache with Redis
- Introduce cdb.ts for cached database connection using Drizzle ORM and RedisCache
- Update imports in gateway and worker to use db cache helpers and cdb
- Remove cache package dependency from cache package and add it to db package
- Update pnpm-lock.yaml and package.json files accordingly

This refactor centralizes caching logic in the db package and integrates Redis caching with Drizzle ORM for improved cache management.
