+++ apps/gateway/package.json
@@ -36,18 +36,13 @@
 		"@opentelemetry/api": "1.9.0",
 		"@opentelemetry/sdk-node": "0.205.0",
 		"dotenv": "16.5.0",
-		"drizzle-orm": "1.0.0-beta.1-57a367f",
 		"gpt-tokenizer": "3.0.1",
 		"hono": "^4.9.7",
 		"hono-openapi": "^0.4.6",
 		"ioredis": "5.6.1",
-		"pg": "^8.16.3",
 		"redis": "5.8.2",
 		"worker": "workspace:*",
 		"zod": "3.25.75",
 		"zod-openapi": "^4.2.4"
-	},
-	"devDependencies": {
-		"@types/pg": "^8.15.5"
 	}
 }
