+++ apps/gateway/src/lib/iam.ts
@@ -1,8 +1,7 @@
 import { HTTPException } from "hono/http-exception";
 
+import { cdb as db } from "@llmgateway/db";
 import { models, type ModelDefinition } from "@llmgateway/models";
-
-import { cachedDb as db } from "./cached-db.js";
 
 export interface IamRule {
 	id: string;
