+++ apps/worker/src/worker.ts
@@ -1,13 +1,10 @@
 import Stripe from "stripe";
 import { z } from "zod";
 
-import {
-	getOrganization,
-	consumeFromQueue,
-	LOG_QUEUE,
-} from "@llmgateway/cache";
+import { consumeFromQueue, LOG_QUEUE } from "@llmgateway/cache";
 import {
 	db,
+	getOrganization,
 	log,
 	organization,
 	eq,
