+++ packages/db/src/index.ts
@@ -1,6 +1,8 @@
 import * as schema from "./schema.js";
 
 export * from "./db.js";
+export * from "./cdb.js";
+export * from "./cache-helpers.js";
 export * from "./schema.js";
 export * from "./types.js";
 export * from "./migrate.js";
